Move getTaskName
